Switch to Github API for fetching team.yml.
It turns out that `git` is not installed in AWS lambdas. There exist
methods for packaging git inside a ruby gem or similar so it can be
installed, but these seem like a huge hack.

We'd previously wanted to avoid using the Github API since Github does
not offer finely grained OAuth permission scopes. Specifically, the only
OAuth scope that offers comparable access to private repo content is the
`repo` scope, which grants full read/write access to all private repos
including issues and pull-requests and other meta-content.

This is a nuisance, but as long as we continue using a machine github
user with only read-only access to repositories, it doesn't make much
difference.

Also fix some minor bugs with sending email.