Add config to tell cloudlib where to find config.
This is a little meta, but we don't want to put the full cloudlib config
in this repo since it contains a lot of configuration specific to
Login.gov's infrastructure that is private and not applicable to an open
source project.

But we do want it to be as easy as possible for Login.gov team members
to get up and running.

So add a default pointer to that repo so cloudlib can download it as
part of `cl lambda init`.